Script used to format the training data
This script compiles all training data into a usable data set that can be used to train a hybrid convolutional neural network. The script also has the section used to create the network structure and train it. There is also a section to evaluate the performance of the trained model.